rpm: reorganize `Info`, fs walking code
